Moved InvocationRegistry.notifyLogic into ResponseHandler
Notifyin the invocations is a concern of the response handler; not of the invocation registry. So
the logic has been moved into the ResponseHandler.
